Complete refactoring that aligns Object instance methods
This change ensures that all API calls made from the `Client#Object` class handle 404s by raising a custom exception. When dor-services-app responds with a 404, this is not an unexpected response when the object does not exist. This is how the API should behave. Making this change lets downstream users reason more confidently when the client returns a response.

This also expands the ability to handle 404s beyond just the `#current_version` call.

Fixes #39