Add implemention of catkey endpoint
Fixes #124

This allows dor-services-client to hit the dor-services-app endpoint that resolves barcodes to catkeys. The endpoint has three separate behaviors:

1. Returns 500 when passed no barcode param or when barcode param is empty;
2. Returns 200 and a plain-text body when barcode param is passed and the lookup succeeds;
3. Returns 200 and an empty plain-text body when barcode param is passed and the lookup fails

This pull request accommodates these three cases, returning UnexpectedResponse for the first case, the catkey in the second case, and NotFoundResponse for the third case.

This pull request also pulls some copypasta code into a superclass.